Revert "5710742254_EX1.11"
This reverts commit 6037172609b230c2da7a48b337ede8ee4702506f.

+++ 5710742254_EX1.11/main.c
@@ -3,9 +3,9 @@
 
 int main()
 {
-    int a;
+    float a;
     printf("Enter integer number :");
-    scanf("%d",&a);
+    scanf("%f",&a);
     printf("Octal number : %o\n",a);
     printf("Hexadecimal number : %x",a);
     return 0;
